Updated packages and moved references on Infrastructure
