+++ src/template/src/Genocs.MicroserviceLight.Template.Application/Genocs.MicroserviceLight.Template.Application.csproj
@@ -7,7 +7,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Genocs.MicroserviceLight.Template.Domain\Genocs.MicroserviceLight.Template.Domain.csproj" />
     <ProjectReference Include="..\Genocs.MicroserviceLight.Template.Shared\Genocs.MicroserviceLight.Template.Shared.csproj" />
-
   </ItemGroup>
 
 </Project>
